requirements.txt
python-docx>=1.1.2
python-pptx>=1.0.2
pypdf>=5.4.0